store -> readonly

+++ src/zarr/storage/common.py
@@ -4,7 +4,7 @@ import json
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
 
-from zarr.abc.store import ByteRangeRequest, Store, StoreAccessMode
+from zarr.abc.store import ByteRangeRequest, Store
 from zarr.core.buffer import Buffer, default_buffer_prototype
 from zarr.core.common import ZARR_JSON, ZARRAY_JSON, ZGROUP_JSON, AccessModeLiteral, ZarrFormat
 from zarr.errors import ContainsArrayAndGroupError, ContainsArrayError, ContainsGroupError
@@ -45,9 +45,12 @@ class StorePath:
         self.store = store
         self.path = path
 
-    async def _init(self, mode: AccessModeLiteral) -> None:
+    @classmethod
+    async def open(
+        cls, store: Store, path: str, mode: AccessModeLiteral | None = None
+    ) -> StorePath:
         """
-        Initialize the StorePath based on the provided mode.
+        Open StorePath based on the provided mode.
 
         * If the mode is 'w-' and the StorePath contains keys, raise a FileExistsError.
         * If the mode is 'w', delete all keys nested within the StorePath
@@ -63,15 +66,26 @@ class StorePath:
         FileExistsError
             If the mode is 'w-' and the store path already exists.
         """
+
+        await store._ensure_open()
+        self = cls(store, path)
+
+        # fastpath if mode is None
+        if mode is None:
+            return self
+
+        if store.readonly and mode != "r":
+            raise ValueError(f"Store is read-only but mode is '{mode}'")
+
         match mode:
             case "w-":
                 if not await self.empty():
                     msg = (
-                      f"{self} is not empty, but `mode` is set to 'w-'."
-                      "Either remove the existing objects in storage,"
-                      "or set `mode` to a value that handles pre-existing objects"
-                      "in storage, like `a` or `w`."
-                      )
+                        f"{self} is not empty, but `mode` is set to 'w-'."
+                        "Either remove the existing objects in storage,"
+                        "or set `mode` to a value that handles pre-existing objects"
+                        "in storage, like `a` or `w`."
+                    )
                     raise FileExistsError(msg)
             case "w":
                 await self.delete_dir()
@@ -80,6 +94,8 @@ class StorePath:
                 pass
             case _:
                 raise ValueError(f"Invalid mode: {mode}")
+
+        return self
 
     async def get(
         self,
@@ -216,7 +232,7 @@ async def make_store_path(
     store_like: StoreLike | None,
     *,
     path: str | None = "",
-    mode: StoreAccessMode | None = None,
+    mode: AccessModeLiteral | None = None,
     storage_options: dict[str, Any] | None = None,
 ) -> StorePath:
     """
@@ -272,45 +288,35 @@ async def make_store_path(
     path_normalized = normalize_path(path)
     assert mode in (None, "r", "r+", "a", "w", "w-")
     if isinstance(store_like, StorePath):
-        if mode is not None and mode != store_like.store.mode:
-            _store = store_like.store.with_mode(mode)
-            await _store._ensure_open()
-            store_like = StorePath(_store, path=store_like.path)
         result = store_like / path_normalized
-    elif isinstance(store_like, Store):
-        if mode is not None and mode != store_like.mode:
-            store_like = store_like.with_mode(mode)
-        await store_like._ensure_open()
-        result = StorePath(store_like, path=path_normalized)
-    elif store_like is None:
-        # mode = "w" is an exception to the default mode = 'r'
-        result = StorePath(await MemoryStore.open(mode=mode or "w"), path=path_normalized)
-    elif isinstance(store_like, Path):
-        result = StorePath(
-            await LocalStore.open(root=store_like, mode=mode or "r"), path=path_normalized
-        )
-    elif isinstance(store_like, str):
-        storage_options = storage_options or {}
-
-        if _is_fsspec_uri(store_like):
-            used_storage_options = True
-            result = StorePath(
-                RemoteStore.from_url(store_like, storage_options=storage_options, mode=mode or "r"),
-                path=path_normalized,
-            )
-        else:
-            result = StorePath(
-                await LocalStore.open(root=Path(store_like), mode=mode or "r"), path=path_normalized
-            )
-    elif isinstance(store_like, dict):
-        # We deliberate only consider dict[str, Buffer] here, and not arbitrary mutable mappings.
-        # By only allowing dictionaries, which are in-memory, we know that MemoryStore appropriate.
-        result = StorePath(
-            await MemoryStore.open(store_dict=store_like, mode=mode or "r"), path=path_normalized
-        )
     else:
-        msg = f"Unsupported type for store_like: '{type(store_like).__name__}'"  # type: ignore[unreachable]
-        raise TypeError(msg)
+        # if mode 'r' was provided, we'll open any new stores as read-only
+        _readonly = mode == "r"
+        if isinstance(store_like, Store):
+            store = store_like
+        elif store_like is None:
+            store = await MemoryStore.open(readonly=_readonly)
+        elif isinstance(store_like, Path):
+            store = await LocalStore.open(root=store_like, readonly=_readonly)
+        elif isinstance(store_like, str):
+            storage_options = storage_options or {}
+
+            if _is_fsspec_uri(store_like):
+                used_storage_options = True
+                store = RemoteStore.from_url(
+                    store_like, storage_options=storage_options, readonly=_readonly
+                )
+            else:
+                store = await LocalStore.open(root=Path(store_like), readonly=_readonly)
+        elif isinstance(store_like, dict):
+            # We deliberate only consider dict[str, Buffer] here, and not arbitrary mutable mappings.
+            # By only allowing dictionaries, which are in-memory, we know that MemoryStore appropriate.
+            store = await MemoryStore.open(store_dict=store_like, readonly=_readonly)
+        else:
+            msg = f"Unsupported type for store_like: '{type(store_like).__name__}'"  # type: ignore[unreachable]
+            raise TypeError(msg)
+
+        result = await StorePath.open(store, path=path_normalized, mode=mode)
 
     if storage_options and not used_storage_options:
         msg = "'storage_options' was provided but unused. 'storage_options' is only used for fsspec filesystem stores."
